description
The batch export system is based on the need to process a large amount of UV-Vis data. First of all, because all its formats are .SP, and the front part has a large number of titles that are useless for processing data, it cannot be imported in Origin in large quantities. Therefore, this batch file processing system is to import a large number of UV-Vis .SP files, and delete the useless part of the data in the file and export them as .csv files to the original folder.